refactor : core/domain running 디렉토링 싱글, 배틀 세분화 #PAR-340
