+++ core/domain/src/main/java/online/partyrun/partyrunapplication/core/domain/running/battle/SaveBattleIdUseCase.kt
@@ -1,4 +1,4 @@
-package online.partyrun.partyrunapplication.core.domain.running
+package online.partyrun.partyrunapplication.core.domain.running.battle
 
 import online.partyrun.partyrunapplication.core.data.repository.BattleRepository
 import javax.inject.Inject
