+++ feature/running/src/main/java/online/partyrun/partyrunapplication/feature/running/battle/BattleContentViewModel.kt
@@ -17,12 +17,12 @@ import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import online.partyrun.partyrunapplication.core.common.result.onFailure
 import online.partyrun.partyrunapplication.core.common.result.onSuccess
-import online.partyrun.partyrunapplication.core.domain.running.BattleStreamUseCase
-import online.partyrun.partyrunapplication.core.domain.running.DisposeSocketResourcesUseCase
-import online.partyrun.partyrunapplication.core.domain.running.GetBattleIdUseCase
-import online.partyrun.partyrunapplication.core.domain.running.GetBattleStatusUseCase
-import online.partyrun.partyrunapplication.core.domain.running.GetUserIdUseCase
-import online.partyrun.partyrunapplication.core.domain.running.SaveBattleIdUseCase
+import online.partyrun.partyrunapplication.core.domain.running.battle.BattleStreamUseCase
+import online.partyrun.partyrunapplication.core.domain.running.battle.DisposeSocketResourcesUseCase
+import online.partyrun.partyrunapplication.core.domain.running.battle.GetBattleIdUseCase
+import online.partyrun.partyrunapplication.core.domain.running.battle.GetBattleStatusUseCase
+import online.partyrun.partyrunapplication.core.domain.running.battle.GetUserIdUseCase
+import online.partyrun.partyrunapplication.core.domain.running.battle.SaveBattleIdUseCase
 import online.partyrun.partyrunapplication.core.model.battle.BattleStatus
 import online.partyrun.partyrunapplication.core.model.running.BattleEvent
 import online.partyrun.partyrunapplication.feature.running.util.distanceToCoordinatesMapper
